refactor: criação de EstoqueMapper e mais refatorações
