--- src/main/java/com/esboco_comix/dao/utils/DAOUtil.java
@@ -1,5 +1,0 @@
-package com.esboco_comix.dao.utils;
-
-public class DAOUtil {
-    
-}
